cosmetic improvement
git-svn-id: http://fs.math.tu-berlin.de:8000/svn/jreality/trunk@4872 6a571d9c-c4e9-67b8-d2a0-cd0d85a3e679

+++ src-audio/de/jreality/audio/DelayPath.java
@@ -242,9 +242,9 @@ public class DelayPath implements SoundPath {
 		
 		double cp = r*Math.cos(phi);
 		double sp = r*Math.sin(phi);
-		p[2] = (float) (cp*Math.cos(theta));
-		p[0] = (float) (cp*Math.sin(theta));
-		p[1] = (float) sp;
+		p[2] = cp*Math.cos(theta);
+		p[0] = cp*Math.sin(theta);
+		p[1] = sp;
 	}
 
 	private void advanceFrame() {
